Changing API timestamp format to "T" separator
In the future we may want to have batch_timestamp as a URL parameter,
in which case a space separator is not acceptable. ISO standard allows
for "T" to be used instead.

Changing the README to suggest the use of the "T" separator in the
JSON to encourage consistency between APIs.

Briefly investigated changing the database to use "T" as well, so this
format is used everywhere. This requires registering an adapter and
converter to sqlite3. The file: /usr/lib/python3.6/sqlite3/dbapi2.py
provides all the groundwork for this change.

There is no real benefit to changing the database format if we continue
to allow TestResult to take timestamp strings with either a space or "T"
separator, so leaving as is for now.